Use face ordering graph for face condition. Display graph.
Following Robert Lang's advice, use a graph to represent the ordering of faces in the folded model.  Create a graph where faces are the vertices. Two color the graph to label front/back faces. Where a front face is adjacent to a back face by a mountain crease, make the front face the source and the back face the target, and vice versa for valley creases.  The face condition becomes there are no internal source or sink vertices.